# (perf) - Improve performance of simplify by removing traversal