This PR handles the type information that is lost due to sequence expression transformation in Simplify plugin. It takes a different approach instead of traversing all of the identifiers for type information. 

### Before

pluginAlias                             time(ms) # visits time/visit(ms)
// Run 1
minify-simplify                         6090.013 278330   0.022
transform-simplify-comparison-operators 41.059   18348    0.002
// Run 2
minify-simplify                         6029.847 278330   0.022
transform-simplify-comparison-operators 46.415   18348    0.003
// Run 3
minify-simplify                         6045.218 278330   0.022
transform-simplify-comparison-operators 44.464   18348    0.002

### After
// 1
minify-simplify                         4617.782 278330   0.017
transform-simplify-comparison-operators 122.181  18348    0.007
// 2
minify-simplify                         4718.766 278330   0.017
transform-simplify-comparison-operators 122.642  18348    0.007
// 3
minify-simplify                         4447.601 278330   0.016
transform-simplify-comparison-operators 127.682  18348    0.007

The type inference is based on the Babel inference. 

+ Only for identifiers we take the different path